+++ herd-view/src/app/bin/model/system.ts
@@ -1,0 +1,6 @@
+import { RedHerdObject } from './base/redherd-common';
+
+export interface System extends RedHerdObject {
+    seed: string;
+    dob: string;
+}