+++ herd-view/src/app/services/base/redherd-web.ts
@@ -14,7 +14,10 @@ export class RedHerdWebContext {
   }
 
   public get Token(): string {
-    return localStorage.getItem(Config.auth_token_store);
+    // Switched to sessionStorage API in order to mitigate the effect of the
+    // browser dirty cache after multiple framework deploy
+    //return localStorage.getItem(Config.auth_token_store);
+    return sessionStorage.getItem(Config.auth_token_store);
   }
 
   public get HttpOptions(): { headers: HttpHeaders } {